requirements.txt
openai-whisper
torch
SpeechRecognition
scipy
webrtcvad-wheels
librosa
langdetect
argostranslate
numpy
gTTS
pygame
pyttsx3
edge-tts
pydub
streamlit
streamlit-webrtc
pandas
